Ændret getdiscountbycode til at retunere en double

+++ Service/Model/ServiceDiscount.cs
@@ -14,6 +14,6 @@ namespace Service.Model
         public string DiscountCode { get; set; }
 
         [DataMember]
-        public int DiscountAmount { get; set; }
+        public double DiscountAmount { get; set; }
     }
 }
